feat: 修复编译器类型处理和枚举编译问题
主要变更：
1. 重构编译器中的类型编译方法，使用compile_type替代as_llvm_type
2. 优化枚举类型的注册和编译流程
3. 修复浮点数类型处理，将f64改为f32
4. 调整引用类型和结构体字段访问的编译逻辑
5. 改进枚举类型的标签和数据字段处理
6. 优化引擎中模块处理和访问者模式的调用方式

+++ src/lexer/mod.rs
@@ -322,7 +322,7 @@ impl Lexer {
         }
         pos.set_span(v.len());
         if is_decimal {
-            let f: f64 = v.parse().unwrap();
+            let f: f32 = v.parse().unwrap();
             return Some((Token::Float(f), pos));
         }
         let i: i64 = v.parse().unwrap();
